+++ packages/fpcast-test/fpcast-test/fpcast-test.c
@@ -31,6 +31,36 @@ set_two(PyObject* self, PyObject* value)
   return 0;
 }
 
+/**
+ * This get-set pair test that the getter/setter call trampolines interact
+ * correctly with stack switching. They are used in `src/tests/test_syncify.py`.
+ * We see that the descriptor trampoline is used because the signatures don't
+ * take a closure argument. We assign `getset_func` to be a function that calls
+ * `syncify()` so that if the trampoline uses JS frames then the stack switch
+ * would fail.
+ */
+PyObject* getset_func = NULL;
+
+static PyObject*
+get_one_call(PyObject* self)
+{
+  return PyObject_CallNoArgs(getset_func);
+}
+
+static int
+set_two_call(PyObject* self, PyObject* value)
+{
+  Py_CLEAR(getset_func);
+  Py_INCREF(value);
+  getset_func = value;
+  if (value == Py_None) {
+    return 0;
+  }
+  PyObject* result = PyObject_CallNoArgs(value);
+  Py_XDECREF(result);
+  return result ? 0 : -1;
+}
+
 // These two structs are the same but it's important that they have to be
 // duplicated here or else we miss test coverage.
 static PyMethodDef Test_Functions[] = {
@@ -72,6 +102,9 @@ static PyMethodDef Test_Methods[] = {
 static PyGetSetDef Test_GetSet[] = {
   { "getset0", .get = (getter)zero },
   { "getset1", .get = (getter)one, .set = (setter)set_two },
+  { "getset_jspi_test",
+    .get = (getter)get_one_call,
+    .set = (setter)set_two_call },
   { NULL }
 };