+++ src/core/stack_switching/stack_switching.mjs
@@ -1,0 +1,22 @@
+/**
+ * Files exported from here are copied into the Emscripten namespace.
+ * See esbuild.config.mjs.
+ */
+
+import {
+  jsWrapperTag,
+  wrapException,
+  adjustWasmImports,
+} from "./create_invokes.mjs";
+import { initSuspenders } from "./suspenders.mjs";
+
+export { promisingApply, createPromising } from "./suspenders.mjs";
+
+export { jsWrapperTag };
+
+Module.preRun.push(initSuspenders);
+
+if (jsWrapperTag) {
+  Module.adjustWasmImports = adjustWasmImports;
+  Module.wrapException = wrapException;
+}